Use Raven's synchronous blocking transport (HTTPTransport)
Raven's default transport is ThreadedHTTPTransport: an "async worker for
processing messages". This is probably overkill for our usage as making
the Raven Client process messages asynchronously in a worker seems to be
unnecessary given cron-sentry's role as a simple CLI wrapper for Sentry.

There's a more frustrating issue though with the ThreadedHTTPTransport:
it has an aggressive initial timeout of 0.1 seconds after which it will
print a warning to stdout with the total timeout it will wait and key
sequences to abort the sending process. In a non-interactive context
this is unhelpful, but worse, it will result in output even when running
cron-sentry with the --quiet parameter to silence output.

If the Sentry server is at all slow to respond this will result in Cron
email alerts being sent even when running with --quiet with just the
"Sentry is attempting to send 1 pending error messages..." as content.

We could potentially try and fix this upstream in various ways, but the
synchronous worker will fix this anyway, and is probably a better fit.